# app/Modules/BuddyClient/Sources/BuddyClient/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/health/ping": {
            "get": {
                "summary": "Ping",
                "operationId": "ping_health_ping_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PingResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/app-api/v1/auth/register": {
            "post": {
                "summary": "Register",
                "operationId": "register_app_api_v1_auth_register_post",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_register_app_api_v1_auth_register_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Return whether the user account has been created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Provided user credentials already exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid payload provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/app-api/v1/auth/login": {
            "post": {
                "summary": "Login",
                "operationId": "login_app_api_v1_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_login_app_api_v1_auth_login_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Return the access token of the user logging in.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid payload provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/app-api/v1/auth/session": {
            "get": {
                "summary": "Session",
                "operationId": "session_app_api_v1_auth_session_get",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the session data for the user that is logged in.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid payload provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/app-api/v1/auth/refresh": {
            "post": {
                "summary": "Refresh",
                "operationId": "refresh_app_api_v1_auth_refresh_post",
                "parameters": [
                    {
                        "name": "refresh-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Refresh-Token"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_login_app_api_v1_auth_login_post": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "title": "Body_login_app_api_v1_auth_login_post"
            },
            "Body_register_app_api_v1_auth_register_post": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "title": "Body_register_app_api_v1_auth_register_post"
            },
            "BuddyErrorDetail": {
                "properties": {
                    "type": {
                        "type": "string",
                        "title": "Type"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Msg"
                    }
                },
                "type": "object",
                "required": [
                    "type",
                    "msg"
                ],
                "title": "BuddyErrorDetail"
            },
            "ErrorResponse": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/BuddyErrorDetail"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ErrorResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "LoginResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "const": "OK",
                        "title": "Detail"
                    },
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "expiry_timestamp": {
                        "type": "integer",
                        "title": "Expiry Timestamp"
                    },
                    "token_type": {
                        "type": "string",
                        "const": "bearer",
                        "title": "Token Type"
                    },
                    "refresh_token": {
                        "type": "string",
                        "title": "Refresh Token"
                    }
                },
                "type": "object",
                "required": [
                    "detail",
                    "access_token",
                    "expiry_timestamp",
                    "token_type",
                    "refresh_token"
                ],
                "title": "LoginResponse"
            },
            "PingResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "const": "PONG",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "PingResponse"
            },
            "RegisterResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "const": "Created",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "RegisterResponse"
            },
            "SessionResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "const": "OK",
                        "title": "Detail"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    }
                },
                "type": "object",
                "required": [
                    "detail",
                    "user"
                ],
                "title": "SessionResponse"
            },
            "UserResponse": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "email"
                ],
                "title": "UserResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}